% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow01.R
\name{incidence_pull}
\alias{incidence_pull}
\alias{dynamic_output_figure}
\alias{incidence_output_figure}
\alias{epiestim_tibble_rt}
\alias{epiestim_current_rt}
\alias{figure_tsibble_rt}
\alias{cdc_datatable_html}
\alias{cdc_dotwhiskers_ggplot}
\alias{nested_figure_01}
\alias{nested_figure_02}
\alias{nested_figure_03}
\alias{nested_figure_04}
\alias{nested_table_01}
\alias{nested_table_02}
\alias{nested_table_03}
\title{MAIN TITLE}
\usage{
incidence_pull(data, variable_date, ...)

dynamic_output_figure(data, fit, date_start, date_end, nombre_area,
  date_breaks = "7 day")

incidence_output_figure(data, date_start, date_end, nombre_area,
  date_breaks = "7 day")

epiestim_tibble_rt(epiestim_output)

epiestim_current_rt(epiestim_tibble_rt)

figure_tsibble_rt(epiestim_tibble_rt, y_breaks = 10,
  date_breaks = "7 day")

cdc_datatable_html(data)

cdc_dotwhiskers_ggplot(data, nest_level = nm_dist,
  high_level = nm_depa, order = n_pos_clean, point = rt_estimate,
  xmax = rt_conf.upper, xmin = rt_conf.lower)

nested_figure_01(data)

nested_figure_02(data)

nested_figure_03(data)

nested_figure_04(data)

nested_table_01(data)

nested_table_02(data)

nested_table_03(data)
}
\arguments{
\item{data}{data}

\item{variable_date}{variable_date}

\item{fit}{fit}

\item{date_start}{date_start}

\item{date_end}{date_end}

\item{nombre_area}{nombre_area}

\item{date_breaks}{date_breaks}

\item{epiestim_output}{epiestim_output}

\item{epiestim_tibble_rt}{epiestim_tibble_rt}

\item{y_breaks}{y_breaks}

\item{nest_level}{nest_level}

\item{high_level}{high_level}

\item{order}{order}

\item{point}{point}

\item{xmax}{xmax}

\item{xmin}{xmin}

\item{date_start}{date_start}

\item{date_end}{date_end}

\item{nombre_area}{nombre_area}
}
\value{
incidence workflow!
}
\description{
main description
}
\section{Functions}{
\itemize{
\item \code{incidence_pull}: specific description

\item \code{dynamic_output_figure}: create figure: incidence original plot number 1

\item \code{incidence_output_figure}: create figure: incidence original plot number 2

\item \code{epiestim_tibble_rt}: Rt estimates: one table with dates and R values

\item \code{epiestim_current_rt}: Rt estimates: filter and select data

\item \code{figure_tsibble_rt}: create figure: time serie of Rt

\item \code{cdc_datatable_html}: other description

\item \code{cdc_dotwhiskers_ggplot}: create figure: dot whiskers ggplot

\item \code{nested_figure_01}: figure 1: incidence plot

\item \code{nested_figure_02}: figure 2: time serie of Rt per strata

\item \code{nested_figure_03}: figure 3: dot-whiskers plot of each strata

\item \code{nested_figure_04}: figure 4: map of Rt per strata

\item \code{nested_table_01}: table 1: growth rate and doubling time. includes halving time if there is a split

\item \code{nested_table_02}: table 2: goodness of fit per strata

\item \code{nested_table_03}: table 3: Rt per strata
}}

