% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow01.R
\name{cdc_dotwhiskers_ggplot}
\alias{cdc_dotwhiskers_ggplot}
\title{MAIN TITLE}
\usage{
cdc_dotwhiskers_ggplot(
  data,
  nest_level = nm_dist,
  high_level = nm_depa,
  order = n_pos_clean,
  point = rt_estimate,
  xmax = rt_conf.upper,
  xmin = rt_conf.lower
)
}
\value{
incidence workflow!
}
\description{
main description
}
\section{Functions}{
\itemize{
\item \code{cdc_dotwhiskers_ggplot}: specific description
}}

